v1: provide deployment for GAE using Docker
The Phoenix app running the Risteys v1 portal can now be deployed to
Google App Engine using the flexible environment with our own Docker
container.

Note that for the app to run correctly (i.e. with websocket support), it
must be hosted on risteys-daly.appspot.com.
If the domain is different, change it in config/prod.exs.