fix js search box error on phenotype page
Add a conditional to check if the "home-search-input" element is
defined on the current page, and only call ".focus()" when this
element exists.

Currently the phenotype page doesn't have the search feature which
resulted in the JS throwing an error.
More importantly, any JS code below the "search_box.focus()" line
would not be run because of the previous error.