risteys: add bin tooltip, use D3 for cumulative
We use D3 to add a tooltip when the mouse hovers over a bin. This
hooks an event on each SVG rect. It appears that if the SVG is part of
a DOM managed by VueJS, then the mouseover events are not set on the
rect. Maybe because VueJS takes over the DOM and hence the SVG.

So adding the tooltip also required to use D3 instead of VueJS to do
the cumulative toggle.

NOTE: This changes introduces an JS output error in the console
"Error: unknown type: mouseover"
but the plot and toggle are still working correctly.