risteys: remove all NaN in plot data
NaN values means individual-level data (between 1 and 5 individuals).
Previously, only NaN values in the tails were removed. This caused the
"toggle cumulative" button do be not applicable, as it would have to
computed X + NaN.
Now, all NaN values are removed from the distribution data, so the
"toggle cumulative" button always work.

As always, if NaN values are removed, we indicate this to the user in
the plot subtitle.