pipeline: remove sex covariate for sex-endpoints
When doing survival analysis using the Cox model we put sex as a
covariate to adjust for sex. Some of the outcome endpoints are
sex-specific (breast cancer for example), so the sex covariate
completely determines the outcome endpoints. In turn, this leads to
ConvergenceWarning in the lifelines library.

The solution added with this commit is to check first if an endpoint
is sex-specific. If it is, then we remove sex from the covariates of
the Cox model. If it is not, then we keep the model as is.