pipeline: fix surv analysis when pair has DEATH
When we introduced the DataFrame trimming previously, we first copied
the KEEP columns and then added prior and outcome columns. The DEATH
column being in KEEP and potentially in prior or outcome would lead to
a df with 2 columns DEATH, which resulted in the following error:
    File "/mnt/data/script/surv_analysis.py", line 390, in set_timeline
      pseudo_later_year = df["birth_year"] + df[later_age]
    ...
    TypeError: ("'<' not supported between instances of 'str' and 'int'", 'occurred at index 403')

This fix removes this possibility, so the DEATH column is never twice
in the df.