risteys: frontend redesign to support mobile
This commit introduces frontend only changes to provide a responsive
layout for support a mobile view of Risteys.

In particular, the CSS has been change to use `@media` query with a
min-width of 650px that defines the mobile breakpoint.
Viewport with 650px-1200px are supported in a "best-effort" way.
Viewport >1200px are supported and well displayed, as before.

The histogram plots have been refactored to support setting a dynamic
width on them. JS has been added to get the window width and display
correctly the histogram plots on mobile.

The survival analyses plot is hidden on mobile.

GitHub: fixes #51 #30