pipeline+risteys: split drug stats (+db migration)
This commit make 2 related changes to drug stats:
1. Split the drug stats table into 2 tables (ATC Drug & Drug Stats)
2. Pipeline output for drug stats is ready to be imported

Previously, a single table was used to store drug stats. In
particular, the drug "name" (or "description") was stored with the
statistics.
We now put the ATC and its description in a separate table, as this
allows us to import the drug statistics without fiddling with merging
the drug description in the same input file.