pipeline: not using longit & remove case-fata.
The endpoint longitudinal file was used previously to compute the
longitudinal metrics (median events, recurrence).
We will not need them anymore, as they will be removed to make space
for the new mortality statistics.

The case-fatality is also removed (could have been cleaner if put in
another commit but oh well) as it will be replaced by mortality done
with survival analysis.