pipeline: use Python script for densify step
At first, we used a Python script with Pandas for the densify step and
it was slow (couple of hours).
Then I made a Rust script to make it faster (2 min).
Now we use a fast Python script (4 min). This to harmonize the
codebase and make it more accessible to everyone, as Rust is probably
less data-scientist friendly than Python.

This Python script is faster than the original one because it doesn't
do any input validation, unlike Pandas in the original script.