maint: bring R3-6 versions into their own dir
Previously we relied on git branches in this one repository to track
the code of the different Risteys R- versions.

Now we leverage the monorepo setup and bring the Elixir/Phoenix code
of each R3–6 version into their own directory.
This makes it easier to apply cross-version changes, as it will be
done in 1 commit.