web: fix SSL issue with hackney on OTP 24
Previously hackney was locked at version 1.15.1 in the mix.lock
file. However this version was not working on newer versions of
OTP (including v24 that we use now).

We ran mix deps.update hackney to get the latest version available to
still satisfies our direct dependencies from mix.exs.

This updates hackney to v1.18.0 (and its dependencies) which works on
OTP 24.

For developers: once you get this commit, do "mix deps.get" to have
the same dependencies as specified in the updated mix.lock file.