pipe: use Feather v2 instead of Parquet
The ongoing -- not yet released -- rewrite of the pipeline is using
the Feather file format for data files.
So here I change the handling of data files from Parquet to Feather,
in order to be consistent throughout the pipeline.

Both of these file formats are good for our use case.
Parquet seems oriented towards long-term storage, due to a more stable
format and higher compression.
Feather v2 seems oriented towards general purpose storage, with
usually faster reads and writes, and is more thightly coupled to the
Arrow ecosystem.