web: improve import of upset plots (+db migr)
While QC-ing Risteys with FinnGen R9.1.0 data it appeared many
endpoints didn't have upset plots or tables. This lead to a confusing
UX since the upset links where showing on the endpoint page
regardless.

This commit adds the status of upset plots and tables to the database
to clarify why an endpoint is missing upset information.

To streamline the import process, this commit integrates the "moving
files around" part to the import script.