web: display "no data" instead of empty elements
Previously, tables or plots would be displayed even if there was no
data to populate them (for example non-core endpoints don't have
mortality statistics).

Now in this commit we remove these elements and display a message
indicating to the user that there is no data to display.
This makes it clearer for the user why there is no data and removes
some visual distraction.