web: rename "phenocode" to "endpoint" (+db migr)
The word "phenocode" was introduced when Risteys started. At the time
it was not clear for me what was the difference between a phenocode
and a FinnGen endpoint, so I just went along with this decision (also
there was already a module named Endpoint as part of Phoenix).
However, as time passes, I realized that phenocode and FinnGen
endpoint were the same thing.
The word phenocode was also confusing for Risteys users as they had
only heard about endpoint.
This commit now replaces all occurences of phenocode to endpoint.

A URL redirect from /phenocode/:name to /endpoint/:name is added in
this commit.
It is important to keep that redirect in order for URLs that were
shared or published to keep working.

The renaming occurs in various places:
- References to "phenocode" on the frontend
- "phenocode" being part of the URL when looking at an endpoint page
- References to "pheno" and "phenocode" on the backend
  - Variable names in /lib/*, in import scripts
  - Module names
    In particular, Risteys.Phenocode is now Risteys.FGEndpoint.Definition,
    this fits better the Phoenix mindset around Context, as well as
    the domain.
  - Database tables and columns
  - Routing / URL

GitHub: closes #118