web: fix page load error for endpoints with 0 case
Previously, the endpoint page would return a HTTP error 500 "Internal
Server Error" for endpoints that don't have any cases.

This was due to computing the relative case count in the endpoint
explainer, which tried to divide by 0.

This commit fixes the issue by effectively showing an empty bar plot
when there are no cases.

GitHub: fixes #94