web: add explainer step with cov pheno filtering
Previously, the last step of the intermediate case counts would not
always reflect the total number of individuals shown in the key
figures. This is because the Risteys pipeline filters out some
individuals based on the "COV PHENO" file, in order to have the same
individuals used in PheWeb.

We now display a last step that reflects that filtering.

GitHub: fixes #123