Allow usage of R8 data, codebase & DB based on R9
FR_FG branch is based on master branch for R9 Risteys. However, FR_FG
branch does not use R9 data, because in FinRegistry, endpoints are only run
using R8 definitions. Changes in this commit allow importing of R8 data to
database that has a new structure according to the R9 data, and makes the.
This way no changes are needed elsewhere in the codebase, because e.g.
database querying works without changes.