HTTPS redirection work on App Engine flexible env
It appeared that https redirection in an app.yaml Google Cloud
configuration file does not work when using Google Cloud App Engine
flexible environment. Risteys has to use a flexible environment, not a
standard environment because it's written in Elixir. Therefore,
force_ssl: [rewrite_on: [:x_forwarded_proto]] configuration is needed in
the prod.exs configuration file to force the application to use the
secure transfer protocol used by HTTPS when HTTP request is sent.

An empty line is added to the end of app.yaml file because GitHub
complained about missing new line at the end of the file.

git closes #162