Update endpoint definitions to be in a table, v1
So far used endpoint definition diagram is a static in a way that
arrows in the diagram break down when anything is changed, so they only
work when everything remains the same.

A table format of endpoint definition introduced in this commit aims to:
1. make the section technically more robust & flexible to allow FR data
2. allows presenting counts clearly from both FinRegistry and FinnGen

This commit only introduces the first version of the table because FR
intermediate counts are not yet available. Code changes to actually get
the FR counts will be added later. New features may be added later too.