Handle long endpoint names in Mortality section
Previously long endpoint names broke the structure of Mortality section
and even caused content to overflow because the names did not wrap.

Now the first occurance of endpoint name in Mortality section is able to
wrap, even though it doesn't contain white spaces. For later occurances
of the name, it's cut at 24 char, if it's longer than that