Scale CIF plot y-axis to content & show values
Previously range of cumulative incidence function (CIF) plot y-axis was
always from 0 to 100. This caused the plot line to be flat for endpoints
with low prevalence, and thus the values of probability of first event
and the shape of the fuction were not (easily) visible.

Now the y-axis is scaled to content, so that the plot always fits the
data. Also the values shown when hovering over the CIF line are now
shown in one decimal point precision (were integers) because the scaled
y-axis enables to see the values more exactly.

git close #199