Fix compilation warnings, no functionality changes
There were unused variable warnings for variables "url", "age",
"max_value", and "mortality_data". Variable "age" is in pattern matching
for a map that has keys "age" and "value" but the age value was not later
used. Solved by refixing age with underscore because it was unused. Other
unused variables were cases where data was saved to a variable at the end
of a function to return it from the function call. In these cases it was
unnecessary to save the data to a function because the value/data that
exists at the end of the function is returned anyway.

Making alias for StatsSex module and importing IO caused warnings because
these modules were not used in the code. Removed the alias and import
lines because those had been just forgotten to remove previously.

Related to commits 7be536a, 9fe2bd2, dd54813, and 91aa4e3.