Removed unused function relative_count/2
Removed definition of relative_count/2 function because it was unused
and thus caused warning in file compilation. The function is unused
because "progress bar" of relative counts was removed in commit 8b667c9
as it was not always showing the true chanage in endpoint case counts in
comparison to start because adding other endoints in endpoint definition
can lead percentage to go over 100.