Handle endpoints using core status (+control defs)
Have information about controls and core status of endpoints.

Handle non-core endpoints the same way as in FG Risteys, show info
banner and point to a core counterpart if exists.

To enable pointing to _EXALLC or _EXMORE counterparts, need to have
the counterpart endpoints. For that,  need to
- import control information of endpoints
- have also FR results for those endpoints (#213, solved in 2023759f)

git resolves #220