web: fix relationship table infinite loading
Previously when a user went to an endpoint page for an endpoint that was
excluded in FinRegistry they would see the spinner and "Loading data".
However, the request to get the data is never made for excluded
endpoint, leaving the user waiting forever.

This commit provides a quick fix.
It shows a useful text to the user that data is not loaded for excluded
endpoint, instead of showing the infinite spinner.

However, a better fix would be to fetch the data an display it, as their
might be data between an excluded endpoint and a non-excluded endpoint.

GitHub: closes #237