web: fix Ueberauth config erroring in prod
Previously, the Ueberauth config was set in the runtime configuration.
It worked in dev because dev.exs initializes the plug at runtime, and
Ueberauth is a plug.
This did not work in prod, has the Ueberauth plug config needs to be
set at compile time. It was failing with a MatchError at compile time.

In this commit, I move the Ueberauth config to config.exs so that it
works both in prod and dev.