web:lab: remove NPeople relative to group column
The initial idea was to show differences in N people within a group of lab
tests. Because the other N people plot is not useful when N is low.

However, that turned into something confusing after feedback from users. Mostly
because it is not clear how the ratio is computed. It would make more sense to
have the total N people in the group as denominator, instead of the max N people
of the group.