web:lab: add lab value distributions in backend
Note about database/backend side: I decided to collect all distributions for
each OMOP concept ID, and store them as structured data within a single value of
the database. This reduces the complexity of database access (both for reading
and writing), at the expense of more work and complexity in the import step.