web:lab: fix bin reconstruct
The current bin reconstruction algorithm could sometimes end up discarding
many bins on the right side of the distribution. This was due to floating-point
number comparison, in the case where the breaks and bins had different precision
and somehow 2 consecutive breaks fit into 1 bin. This resulted in all the
remaining bins being discarded.

Now, the bin reconstruction is based on the bins rather than on the breaks, to
make sure no information from the bins is discarded.