web: remove rel.table extremity from codebase
This feature has been removed in the frontend for some time because of
performance issues. But it was not completely removed from the codebase, leading
to noise in the application logs.

Adding this feature again would be doable by pre-computing the extremities and
storing them in the database.


Related git commit: 7d949643